third_party/blink/web_tests/http/tests/credentialmanager/credentialscontainer-create-with-virtual-authenticator.html

<!DOCTYPE html>
<title>Credential Manager: End-to-end tests for create() with a virtual authenticator.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings_lite.js"></script>
<script src="/gen/mojo/public/mojom/base/time.mojom-lite.js"></script>
<script src="/gen/url/mojom/url.mojom-lite.js"></script>
<script src="/gen/third_party/blink/public/mojom/webauthn/authenticator.mojom-lite.js"></script>
<script src="/gen/third_party/blink/public/mojom/webauthn/virtual_authenticator.mojom-lite.js"></script>
<script src="resources/test-inputs.js"></script>
<script src="resources/virtual-navigator-credentials.js"></script>
<body>
<script>

if (document.location.host != "subdomain.example.test:8443") {
  document.location = "https://subdomain.example.test:8443/credentialmanager/credentialscontainer-create-with-virtual-authenticator.html";
  promise_test(_ => new Promise(_ => {}), "Stall tests on the wrong host.");
}

authenticatorSetup("CTAP2_0", () => {
  promise_test(async t => {
    let authenticators = await navigator.credentials.test.authenticators();
    assert_equals(authenticators.length, 1);
    let testAuthenticator = authenticators[0];
    assert_true(await testAuthenticator.generateAndRegisterKey(ACCEPTABLE_CREDENTIAL_ID, "subdomain.example.test"));

    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.excludeCredentials = [ACCEPTABLE_CREDENTIAL];
    return promise_rejects_dom(t, "InvalidStateError",
      navigator.credentials.create({ publicKey : customMakeCredOptions}));
  }, "navigator.credentials.create() re-registration returns InvalidStateError");

  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.rp.id = "";
    return promise_rejects_dom(t, "SecurityError",
      navigator.credentials.create({ publicKey : customMakeCredOptions}));
  }, "navigator.credentials.create() with empty rpId returns SecurityError");

  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.authenticatorSelection.requireResidentKey = true;
    return navigator.credentials.create({ publicKey : customMakeCredOptions });
  }, "navigator.credentials.create() with requireResidentKey is supported");

  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.authenticatorSelection.requireResidentKey = false;
    customMakeCredOptions.authenticatorSelection.residentKey = "required";
    return navigator.credentials.create({ publicKey : customMakeCredOptions });
  }, "navigator.credentials.create() resident_key overrides resident_key_required");

  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.authenticatorSelection.residentKey = "required";
    return navigator.credentials.create({ publicKey : customMakeCredOptions });
  }, "navigator.credentials.create() with resident_key is supported");

  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.authenticatorSelection.resident_key = "foobarbaz";
    return navigator.credentials.create({ publicKey : customMakeCredOptions });
  }, "navigator.credentials.create() invalid resident_key value is ignored");

  promise_test(async t => {
    authAbortController = new AbortController();
    authAbortSignal = authAbortController.signal;
    return navigator.credentials.create({ publicKey : MAKE_CREDENTIAL_OPTIONS, signal : authAbortSignal});
  }, "navigator.credentials.create() with abort signal flag not set");

  promise_test(async t => {
    authAbortController = new AbortController();
    authAbortSignal = authAbortController.signal;
    authAbortController.abort();
    var promise = navigator.credentials.create({ publicKey : MAKE_CREDENTIAL_OPTIONS, signal : authAbortSignal});
    return promise_rejects_dom(t, "AbortError", promise);
  }, "navigator.credentials.create() with abort signal flag set before sending request");

  promise_test(async t => {
    authAbortController = new AbortController();
    authAbortSignal = authAbortController.signal;
    var promise = navigator.credentials.create({ publicKey : MAKE_CREDENTIAL_OPTIONS, signal : authAbortSignal});
    authAbortController.abort();
    return promise_rejects_dom(t, "AbortError", promise);
  }, "navigator.credentials.create() with abort signal flag is set after sending request");

  promise_test(async t => {
    authAbortController = new AbortController();
    authAbortSignal = authAbortController.signal;
    let promise = await navigator.credentials.create({ publicKey : MAKE_CREDENTIAL_OPTIONS, signal : authAbortSignal});
    authAbortController.abort();
    return promise;
  }, "navigator.credentials.create() with abort signal flag is set after promise resolved");

  // TODO(nsatragno): move largeBlob tests to WPTs once the largeBlob webdriver
  // extension is implemented.
  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.authenticatorSelection.requireResidentKey = true;
    customMakeCredOptions.extensions = {
      largeBlob: {
        read: true,
      },
    };
    return promise_rejects_dom(t, "NotSupportedError",
      navigator.credentials.create({ publicKey : customMakeCredOptions }));
  }, "navigator.credentials.create() with largeBlob.read set");

  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.authenticatorSelection.requireResidentKey = true;
    customMakeCredOptions.extensions = {
      largeBlob: {
        write: new ArrayBuffer(),
      },
    };
    return promise_rejects_dom(t, "NotSupportedError",
      navigator.credentials.create({ publicKey : customMakeCredOptions }));
  }, "navigator.credentials.create() with largeBlob.write set");

  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.authenticatorSelection.requireResidentKey = true;
    customMakeCredOptions.extensions = {
      largeBlob: {
        support: "preferred",
      },
    };
    var credential =
      await navigator.credentials.create({ publicKey : customMakeCredOptions });
    assert_false(credential.getClientExtensionResults().largeBlob.supported);
    assert_not_own_property(credential.getClientExtensionResults().largeBlob, "blob");
    assert_not_own_property(credential.getClientExtensionResults().largeBlob, "written");

  }, "navigator.credentials.create() with largeBlob.support set to preferred and not supported by authenticator");

  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.authenticatorSelection.requireResidentKey = true;
    customMakeCredOptions.extensions = {
      largeBlob: {},
    };
    // Should be equivalent to preferred.
    var credential =
      await navigator.credentials.create({ publicKey : customMakeCredOptions });
    assert_false(credential.getClientExtensionResults().largeBlob.supported);
    assert_not_own_property(credential.getClientExtensionResults().largeBlob, "blob");
    assert_not_own_property(credential.getClientExtensionResults().largeBlob, "written");

  }, "navigator.credentials.create() with largeBlob.support not set and not supported by authenticator");

  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.authenticatorSelection.requireResidentKey = true;
    customMakeCredOptions.extensions = {
      largeBlob: {
        support: "required",
      },
    };
    return promise_rejects_dom(t, "NotAllowedError",
      navigator.credentials.create({ publicKey : customMakeCredOptions }));
  }, "navigator.credentials.create() with largeBlob.support set to required and not supported by authenticator");
}, {});

authenticatorSetup("With large blob", () => {
  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.authenticatorSelection.requireResidentKey = true;
    customMakeCredOptions.extensions = {
      largeBlob: {
        support: "preferred",
      },
    };
    var credential =
      await navigator.credentials.create({ publicKey : customMakeCredOptions });
    assert_true(credential.getClientExtensionResults().largeBlob.supported);
    assert_not_own_property(credential.getClientExtensionResults().largeBlob, "blob");
    assert_not_own_property(credential.getClientExtensionResults().largeBlob, "written");
  }, "navigator.credentials.create() with largeBlob.support set to preferred and supported by authenticator");

  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.authenticatorSelection.requireResidentKey = true;
    customMakeCredOptions.extensions = {
      largeBlob: {},
    };
    // Should be equivalent to preferred.
    var credential =
      await navigator.credentials.create({ publicKey : customMakeCredOptions });
    assert_true(credential.getClientExtensionResults().largeBlob.supported);
    assert_not_own_property(credential.getClientExtensionResults().largeBlob, "blob");
    assert_not_own_property(credential.getClientExtensionResults().largeBlob, "written");
  }, "navigator.credentials.create() with largeBlob.support not set and supported by authenticator");

  promise_test(async t => {
    var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
    customMakeCredOptions.authenticatorSelection.requireResidentKey = true;
    customMakeCredOptions.extensions = {
      largeBlob: {
        support: "required",
      },
    };
    var credential =
      await navigator.credentials.create({ publicKey : customMakeCredOptions });
    assert_true(credential.getClientExtensionResults().largeBlob.supported);
    assert_not_own_property(credential.getClientExtensionResults().largeBlob, "blob");
    assert_not_own_property(credential.getClientExtensionResults().largeBlob, "written");
  }, "navigator.credentials.create() with largeBlob.support set to required and supported by authenticator");
}, {
  protocol: blink.test.mojom.ClientToAuthenticatorProtocol.CTAP2,
  ctap2version: blink.test.mojom.ClientToAuthenticatorProtocol.CTAP2_1,
  hasLargeBlob: true,
});

</script>
